# openapi/v2.yaml
openapi: 3.0.3
info:
  title: Santaizi HTTP API
  version: 2.0.0
  description: |
    Santaizi Primary 的公开状态和管理接口。浏览器会话使用 Cookie 与 CSRF，
    自动化客户端使用 Bearer Token。Telemetry/Replication 可靠链路仍使用 gRPC。
servers:
- url: /
  description: 当前站点（同源）
- url: https://monitor.example.com
  description: 示例部署地址
...
security:
- cookieAuth: []
- bearerAuth: []
paths:
...
  /api/v2/admin/servers:
    get:
      tags:
      - Servers
      summary: 服务器列表
      description: 管理服务器列表。`host.ip` / `host.ipv4` / `host.ipv6` 为探针上报的明文公网地址，不脱敏。
      operationId: listServers
      parameters:
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/PageSize'
      - $ref: '#/components/parameters/Search'
      - $ref: '#/components/parameters/Sort'
      - $ref: '#/components/parameters/Order'
      responses:
        '200':
          $ref: '#/components/responses/ServerListResponse'
        default:
          $ref: '#/components/responses/Problem'
    post:
      tags:
      - Servers
      summary: 创建服务器
      operationId: createServer
      x-ui-route: /admin/servers
      security:
      - cookieAuth: []
        csrfHeader: []
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        $ref: '#/components/requestBodies/ServerWrite'
      responses:
        '201':
          $ref: '#/components/responses/ServerResponse'
        default:
          $ref: '#/components/responses/Problem'
...
  /api/v2/admin/notifications:
    get:
      tags:
      - Notifications
      summary: 通知渠道列表
      operationId: listNotifications
      parameters:
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/PageSize'
      - $ref: '#/components/parameters/Search'
      - $ref: '#/components/parameters/Sort'
      - $ref: '#/components/parameters/Order'
      responses:
        '200':
          $ref: '#/components/responses/NotificationChannelListResponse'
        default:
          $ref: '#/components/responses/Problem'
    post:
      tags:
      - Notifications
      summary: 创建通知渠道
      operationId: createNotification
      x-ui-route: /admin/notifications
      security:
      - cookieAuth: []
        csrfHeader: []
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        $ref: '#/components/requestBodies/NotificationChannelWrite'
      responses:
        '201':
          $ref: '#/components/responses/NotificationChannelResponse'
        default:
          $ref: '#/components/responses/Problem'
...
  /api/v2/admin/ddns:
    get:
      tags:
      - Network
      summary: DDNS 配置列表
      operationId: listDDNSProfiles
      parameters:
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/PageSize'
      - $ref: '#/components/parameters/Search'
      - $ref: '#/components/parameters/Sort'
      - $ref: '#/components/parameters/Order'
      responses:
        '200':
          $ref: '#/components/responses/DDNSProfileListResponse'
        default:
          $ref: '#/components/responses/Problem'
    post:
      tags:
      - Network
      summary: 创建 DDNS 配置
      operationId: createDDNSProfile
      x-ui-route: /admin/ddns
      security:
      - cookieAuth: []
        csrfHeader: []
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        $ref: '#/components/requestBodies/DDNSProfileWrite'
      responses:
        '201':
          $ref: '#/components/responses/DDNSProfileResponse'
        default:
          $ref: '#/components/responses/Problem'
...
  /api/v2/admin/nat:
    get:
      tags:
      - Network
      summary: NAT 隧道列表
      operationId: listNATTunnels
      parameters:
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/PageSize'
      - $ref: '#/components/parameters/Search'
      - $ref: '#/components/parameters/Sort'
      - $ref: '#/components/parameters/Order'
      responses:
        '200':
          $ref: '#/components/responses/NATTunnelListResponse'
        default:
          $ref: '#/components/responses/Problem'
    post:
      tags:
      - Network
      summary: 创建 NAT 隧道
      operationId: createNATTunnel
      x-ui-route: /admin/nat
      security:
      - cookieAuth: []
        csrfHeader: []
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        $ref: '#/components/requestBodies/NATTunnelWrite'
      responses:
        '201':
          $ref: '#/components/responses/NATTunnelResponse'
        default:
          $ref: '#/components/responses/Problem'
...
  /api/v2/admin/settings:
    get:
      tags:
      - Settings
      summary: 获取站点设置
      description: |
        返回当前站点设置。网络相关字段包括 `grpc_host`（公网地址）、
        `proxy_grpc_port`（公网 gRPC 端口，0 表示使用监听端口 `grpcport`）、
        `tls`（探针安装命令是否附加 `--tls` / `-Tls`）。
      operationId: getSettings
      responses:
        '200':
          $ref: '#/components/responses/ObjectResponse'
        default:
          $ref: '#/components/responses/Problem'
    patch:
      tags:
      - Settings
      summary: 更新站点设置
      description: |
        更新并保存站点设置。可写入 `grpc_host`、`proxy_grpc_port`（0–65535，0 表示使用监听端口）
        和 `tls`。这两项用于生成探针/从端安装命令中的公网端口与 TLS 标志，不改变本机 gRPC 监听端口。
      operationId: updateSettings
      x-ui-route: /admin/settings
      security:
      - cookieAuth: []
        csrfHeader: []
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        $ref: '#/components/requestBodies/GenericObject'
      responses:
        '200':
          $ref: '#/components/responses/ObjectResponse'
        default:
          $ref: '#/components/responses/Problem'
  /api/v2/admin/api-tokens:
    get:
      tags:
      - Settings
      summary: API Token 列表
      operationId: listApiTokens
      responses:
        '200':
          $ref: '#/components/responses/APITokenListResponse'
        default:
          $ref: '#/components/responses/Problem'
    post:
      tags:
      - Settings
      summary: 创建 API Token
      operationId: createApiToken
      x-ui-route: /admin/api-tokens
      security:
      - cookieAuth: []
        csrfHeader: []
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        $ref: '#/components/requestBodies/APITokenWrite'
      responses:
        '201':
          $ref: '#/components/responses/APITokenResponse'
        default:
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: santaizi-dashboard
      description: 管理员 OAuth 会话 Cookie
    bearerAuth:
      type: http
      scheme: bearer
      description: 'API Token（Authorization: Bearer）'
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: 浏览器 Cookie 会话写操作必填；值来自 GET /api/v2/auth/session 的 csrf_token
  parameters:
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Search:
      name: q
      in: query
      schema:
        type: string
        maxLength: 200
    Sort:
      name: sort
      in: query
      schema:
        type: string
        maxLength: 64
    Order:
      name: order
      in: query
      schema:
        type: string
        enum:
        - asc
        - desc
        default: desc
...
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: false
      description: Cookie 会话写操作时必填；Bearer Token 调用可省略
      schema:
        type: string
  schemas:
    Problem:
      type: object
      required:
      - title
      - status
      - code
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        code:
          type: string
        detail:
          type: string
        trace_id:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
    Meta:
      type: object
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
          format: int64
        next_cursor:
          type: string
...
    TelemetryPresentation:
      type: object
      properties:
        host:
          type: string
          enum:
          - online
          - offline
          - unknown
        connectivity:
          type: string
          enum:
          - full
          - partial
          - unavailable
          - unknown
        available:
          type: boolean
          nullable: true
        coverage:
          type: string
    PublicNote:
      type: object
      additionalProperties: true
      properties:
        billingDataMod:
          $ref: '#/components/schemas/PublicNoteBilling'
        planDataMod:
          $ref: '#/components/schemas/PublicNotePlan'
        customData:
          $ref: '#/components/schemas/PublicNotePresentation'
    PublicNoteBilling:
      type: object
      additionalProperties: true
      properties:
        startDate:
          type: string
        endDate:
          type: string
          description: Use 0000-00-00 for an unlimited term.
        autoRenewal:
          type: string
          enum:
          - '0'
          - '1'
        cycle:
          type: string
        amount:
          type: string
    PublicNotePlan:
      type: object
      additionalProperties: true
      properties:
        bandwidth:
          type: string
        trafficVol:
          type: string
        trafficType:
          type: string
        IPv4:
          type: string
          enum:
          - '0'
          - '1'
        IPv6:
          type: string
          enum:
          - '0'
          - '1'
        networkRoute:
          type: string
        extra:
          type: string
    PublicNotePresentation:
      type: object
      additionalProperties: true
      properties:
        location:
          type: string
        flag:
          type: string
        orderLink:
          type: string
        buyBtnText:
          type: string
        buyBtnIcon:
          type: string
        slogan:
          type: string
        lat:
          type: string
        lng:
          type: string
        latlng:
          type: string
        locationLabel:
          type: string
    MonitoringOptions:
      type: object
      additionalProperties: false
      properties:
        cpu:
          type: boolean
        memory:
          type: boolean
        disk:
          type: boolean
        network:
          type: boolean
        connections:
          type: boolean
        processes:
          type: boolean
        temperature:
          type: boolean
        gpu:
          type: boolean
        host_info:
          type: boolean
        ip_report:
          type: boolean
        http_probe:
          type: boolean
        icmp_probe:
          type: boolean
        tcp_probe:
          type: boolean
        nat:
          type: boolean
    ServerHost:
      type: object
      description: |
        探针 Host 快照。管理接口以及带 Bearer Token 的公开服务器接口返回明文公网地址（`ip` / `ipv4` / `ipv6`），不脱敏。
        匿名或查看密码访问的 `listPublicServers`、`getPublicServer` 与 `/ws/v2/public/runtime` 省略这些字段。
        其余字段为探针上报的 PascalCase 快照（如 Platform、CountryCode）。
      additionalProperties: true
      properties:
        ip:
          type: string
          description: 探针上报的公网地址；双栈为 `IPv4/IPv6`。管理接口与 Bearer Token 公开请求返回。
        ipv4:
          type: string
          description: 从 `ip` 拆出的 IPv4。管理接口与 Bearer Token 公开请求返回。
        ipv6:
          type: string
          description: 从 `ip` 拆出的 IPv6。管理接口与 Bearer Token 公开请求返回。
    Server:
      type: object
      required:
      - id
      - name
      - tag
      - display_index
      - hide_for_guest
      - enable_ddns
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        tag:
          type: string
        secret:
          type: string
        note:
          type: string
        public_note:
          $ref: '#/components/schemas/PublicNote'
        monitoring_options:
          $ref: '#/components/schemas/MonitoringOptions'
        display_index:
          type: integer
        hide_for_guest:
          type: boolean
        enable_ddns:
          type: boolean
        ddns_profiles:
          type: array
          items:
            type: integer
            format: int64
        host:
          $ref: '#/components/schemas/ServerHost'
        state:
          type: object
          additionalProperties: true
        last_active:
          type: string
          format: date-time
        online:
          type: boolean
        telemetry:
          $ref: '#/components/schemas/TelemetryPresentation'
    ServerWrite:
      type: object
      required:
      - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        tag:
          type: string
          maxLength: 100
        note:
          type: string
          maxLength: 2000
        public_note:
          $ref: '#/components/schemas/PublicNote'
        monitoring_options:
          $ref: '#/components/schemas/MonitoringOptions'
        display_index:
          type: integer
        hide_for_guest:
          type: boolean
        enable_ddns:
          type: boolean
        ddns_profiles:
          type: array
          items:
            type: integer
            format: int64
...
    KeyValue:
      type: object
      required:
      - key
      - value
      properties:
        key:
          type: string
        value:
          type: string
    NotificationChannelWrite:
      type: object
      required:
      - name
      - tag
      - url
      - method
      - request_type
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        tag:
          type: string
          minLength: 1
          maxLength: 100
        url:
          type: string
          format: uri
        method:
          type: string
          enum:
          - get
          - post
        request_type:
          type: string
          enum:
          - json
          - form
        headers:
          type: array
          items:
            $ref: '#/components/schemas/KeyValue'
        body:
          type: string
        verify_tls:
          type: boolean
    NotificationChannel:
      allOf:
      - $ref: '#/components/schemas/NotificationChannelWrite'
      - type: object
        required:
        - id
        properties:
          id:
            type: integer
            format: int64
          created_at:
            type: string
            format: date-time
          updated_at:
            type: string
            format: date-time
...
    DDNSProfileWrite:
      type: object
      required:
      - name
      - provider
      - enable_ipv4
      - enable_ipv6
      - domains
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        provider:
          type: string
        enable_ipv4:
          type: boolean
        enable_ipv6:
          type: boolean
        max_retries:
          type: integer
          format: int64
          minimum: 1
          maximum: 10
        domains:
          type: array
          minItems: 1
          items:
            type: string
        access_id:
          type: string
        access_secret:
          type: string
          format: password
        webhook_url:
          type: string
        webhook_method:
          type: string
          enum:
          - get
          - post
        webhook_request_type:
          type: string
          enum:
          - json
          - form
        webhook_body:
          type: string
        webhook_headers:
          type: array
          items:
            $ref: '#/components/schemas/KeyValue'
...
    DDNSProfile:
      allOf:
      - $ref: '#/components/schemas/DDNSProfileWrite'
      - type: object
        required:
        - id
        properties:
          id:
            type: integer
            format: int64
          created_at:
            type: string
            format: date-time
          updated_at:
            type: string
            format: date-time
    NATTunnelWrite:
      type: object
      required:
      - name
      - server_id
      - target
      - domain
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        server_id:
          type: integer
          format: int64
          minimum: 1
        target:
          type: string
          pattern: ^.+:[0-9]+$
        domain:
          type: string
    NATTunnel:
      allOf:
      - $ref: '#/components/schemas/NATTunnelWrite'
      - type: object
        required:
        - id
        properties:
          id:
            type: integer
            format: int64
          created_at:
            type: string
            format: date-time
          updated_at:
            type: string
            format: date-time
...
    APITokenWrite:
      type: object
      required:
      - note
      - permission
      properties:
        note:
          type: string
          minLength: 1
          maxLength: 200
        permission:
          type: string
          enum:
          - read
          - write
        expires_at:
          type: string
          format: date-time
          nullable: true
          description: 到期时间；省略或 null 表示永不过期
...
    APIToken:
      type: object
      required:
      - id
      - note
      - permission
      - enabled
      - created_at
      properties:
        id:
          type: integer
          format: int64
        note:
          type: string
        permission:
          type: string
          enum:
          - read
          - write
        expires_at:
          type: string
          format: date-time
          nullable: true
        enabled:
          type: boolean
        expired:
          type: boolean
        token:
          type: string
        token_prefix:
          type: string
        created_at:
          type: string
          format: date-time
...
    GenericObject:
      type: object
      additionalProperties: true
    ObjectData:
      type: object
      required:
      - data
      properties:
        data:
          $ref: '#/components/schemas/GenericObject'
...
  requestBodies:
    GenericObject:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenericObject'
    ServerWrite:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ServerWrite'
...
    NotificationChannelWrite:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/NotificationChannelWrite'
...
    DDNSProfileWrite:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/DDNSProfileWrite'
    NATTunnelWrite:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/NATTunnelWrite'
...
    APITokenWrite:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APITokenWrite'
...
  responses:
    Problem:
      description: Problem details
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    ObjectResponse:
      description: Object response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ObjectData'
...
    ServerResponse:
      description: Server
      content:
        application/json:
          schema:
            type: object
            required:
            - data
            properties:
              data:
                $ref: '#/components/schemas/Server'
    ServerListResponse:
      description: Server list
      content:
        application/json:
          schema:
            type: object
            required:
            - data
            - meta
            properties:
              data:
                type: array
                items:
                  $ref: '#/components/schemas/Server'
              meta:
                $ref: '#/components/schemas/Meta'
...
    NotificationChannelResponse:
      description: Notification channel
      content:
        application/json:
          schema:
            type: object
            required:
            - data
            properties:
              data:
                $ref: '#/components/schemas/NotificationChannel'
    NotificationChannelListResponse:
      description: Notification channel list
      content:
        application/json:
          schema:
            type: object
            required:
            - data
            - meta
            properties:
              data:
                type: array
                items:
                  $ref: '#/components/schemas/NotificationChannel'
              meta:
                $ref: '#/components/schemas/Meta'
...
    DDNSProfileResponse:
      description: DDNS profile
      content:
        application/json:
          schema:
            type: object
            required:
            - data
            properties:
              data:
                $ref: '#/components/schemas/DDNSProfile'
...
    DDNSProfileListResponse:
      description: DDNS profile list
      content:
        application/json:
          schema:
            type: object
            required:
            - data
            - meta
            properties:
              data:
                type: array
                items:
                  $ref: '#/components/schemas/DDNSProfile'
              meta:
                $ref: '#/components/schemas/Meta'
    NATTunnelResponse:
      description: NAT tunnel
      content:
        application/json:
          schema:
            type: object
            required:
            - data
            properties:
              data:
                $ref: '#/components/schemas/NATTunnel'
    NATTunnelListResponse:
      description: NAT tunnel list
      content:
        application/json:
          schema:
            type: object
            required:
            - data
            - meta
            properties:
              data:
                type: array
                items:
                  $ref: '#/components/schemas/NATTunnel'
              meta:
                $ref: '#/components/schemas/Meta'
...
    APITokenResponse:
      description: API token
      content:
        application/json:
          schema:
            type: object
            required:
            - data
            properties:
              data:
                $ref: '#/components/schemas/APIToken'
    APITokenListResponse:
      description: API token list
      content:
        application/json:
          schema:
            type: object
            required:
            - data
            - meta
            properties:
              data:
                type: array
                items:
                  $ref: '#/components/schemas/APIToken'
              meta:
                $ref: '#/components/schemas/Meta'
...