# web/packages/api/src/generated/model/serverHost.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Santaizi HTTP API
 * Santaizi Primary 的公开状态和管理接口。浏览器会话使用 Cookie 与 CSRF，
 * 自动化客户端使用 Bearer Token。Telemetry/Replication 可靠链路仍使用 gRPC。
 *
 * OpenAPI spec version: 2.0.0
 */

/**
 * 探针 Host 快照。管理接口以及带 Bearer Token 的公开服务器接口返回明文公网地址（`ip` / `ipv4` / `ipv6`），不脱敏。
 * 匿名或查看密码访问的 `listPublicServers`、`getPublicServer` 与 `/ws/v2/public/runtime` 省略这些字段。
 * 其余字段为探针上报的 PascalCase 快照（如 Platform、CountryCode）。
 */
export interface ServerHost {
  /** 探针上报的公网地址；双栈为 `IPv4/IPv6`。管理接口与 Bearer Token 公开请求返回。 */
  ip?: string;
  /** 从 `ip` 拆出的 IPv4。管理接口与 Bearer Token 公开请求返回。 */
  ipv4?: string;
  /** 从 `ip` 拆出的 IPv6。管理接口与 Bearer Token 公开请求返回。 */
  ipv6?: string;
  [key: string]: unknown;
 }